requirements.txt
azure_storage==0.37.0
elevenlabs==0.2.8
keyboard==0.13.5
mutagen==1.46.0
obs_websocket_py==1.0
openai==1.7.2
pygame==2.3.0
pygame_ce==2.3.0
rich==13.7.0
soundfile==0.12.1
tiktoken==0.5.1
